+++ tket/src/Circuit/CircPool.cpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "CircPool.hpp"
+#include "Circuit/CircPool.hpp"
 
 #include <tkassert/Assert.hpp>
 
-#include "CircUtils.hpp"
-#include "Circuit.hpp"
+#include "Circuit/CircUtils.hpp"
+#include "Circuit/Circuit.hpp"
 #include "Gate/Rotation.hpp"
 #include "OpType/OpType.hpp"
 #include "Utils/Expression.hpp"