+++ tket/src/Circuit/Conditional.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "Conditional.hpp"
+#include "Circuit/Conditional.hpp"
 
 #include "OpType/OpType.hpp"