+++ tket/src/Converters/PhasePoly.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "PhasePoly.hpp"
+#include "Converters/PhasePoly.hpp"
 
 #include <algorithm>
 #include <stdexcept>