+++ tket/src/Gate/GateUnitaryMatrixComposites.cpp
@@ -14,8 +14,8 @@
 
 #include <utility>
 
-#include "GateUnitaryMatrixImplementations.hpp"
-#include "GateUnitaryMatrixUtils.hpp"
+#include "Gate/GateUnitaryMatrixImplementations.hpp"
+#include "Gate/GateUnitaryMatrixUtils.hpp"
 #include "Utils/Constants.hpp"
 
 // This file is for gates with unitary matrices