+++ tket/src/OpType/OpTypeFunctions.cpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "OpTypeFunctions.hpp"
+#include "OpType/OpTypeFunctions.hpp"
 
 #include <memory>
 
-#include "OpType.hpp"
+#include "OpType/OpType.hpp"
 
 namespace tket {
 
@@ -177,6 +177,7 @@ bool is_box_type(OpType optype) {
       OpType::MultiplexedRotationBox,
       OpType::MultiplexedU2Box,
       OpType::StatePreparationBox,
+      OpType::DiagonalBox,
       OpType::ClassicalExpBox,
       OpType::ProjectorAssertionBox,
       OpType::StabiliserAssertionBox,